refactor(cli): split ide.rs into an ide/ module directory
ide.rs (2529 lines) becomes ide/{merge,catalogue,picker,run}.rs with tests
mirrored under ide/tests/. merge = config-file merge logic; catalogue =
client specs/paths; picker = the interactive TUI; run = command entry +
orchestration + write. mod.rs re-exports cmd_ide/cmd_install so
commands::ide::* paths are unchanged. Cross-module helpers bumped to
pub(super). No prod logic or test changed (71 tests preserved).
